+++ src/compile.rs
@@ -361,6 +361,7 @@ impl Uiua {
                 }
             }
             Word::Func(func) => self.func(func, word.span)?,
+            Word::Ocean(prims) => self.ocean(prims, call)?,
             Word::Primitive(p) => self.primitive(p, word.span, call)?,
             Word::Modified(m) => self.modified(*m, call)?,
             Word::Spaces | Word::Comment(_) => {}
@@ -385,6 +386,28 @@ impl Uiua {
             }
         } else {
             return Err(span.sp(format!("Unknown identifier `{ident}`")).into());
+        }
+        Ok(())
+    }
+    fn ocean(&mut self, prims: Vec<Sp<Primitive>>, call: bool) -> UiuaResult {
+        if call {
+            for prim in prims.into_iter().rev() {
+                self.primitive(prim.value, prim.span, true)?;
+            }
+        } else {
+            self.new_functions.push(Vec::new());
+            let span = prims
+                .first()
+                .unwrap()
+                .span
+                .clone()
+                .merge(prims.last().unwrap().span.clone());
+            for prim in prims.into_iter().rev() {
+                self.primitive(prim.value, prim.span, true)?;
+            }
+            let instrs = self.new_functions.pop().unwrap();
+            let function = Function::new(FunctionId::Anonymous(span), instrs, Signature::new(1, 1));
+            self.push_instr(Instr::push(function));
         }
         Ok(())
     }