+++ src/lex.rs
@@ -482,7 +482,9 @@ impl Lexer {
                     let begin = self.loc;
                     self.end(Primitive::Level, start);
                     self.loc = begin;
-                    self.end(Primitive::Left, start);
+                    self.end(Primitive::Water, start);
+                    self.loc = begin;
+                    self.end(Primitive::Rock, start);
                 }
 
                 '(' => self.end(OpenParen, start),